backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.10.0
Pillow>=8.0.0
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
facenet-pytorch>=2.5.2
opencv-python-headless>=4.5.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0 